refactor: decompose dispatcher tests.rs into cohesive submodules
Split the monolithic tests.rs (61 functions, 11+ responsibilities) into
a tests/ subdirectory with focused submodules:

- approval.rs: Tool approval logic and requirements
- auth.rs: Auth detection for tool_auth/tool_activate
- chat_tool.rs: Chat tool standalone execution
- compaction.rs: Message compaction for context recovery
- error_format.rs: Error formatting and text stripping
- image_sentinel.rs: Image data URL handling
- loop_guard.rs: Loop termination and iteration bounds
- pending_approval.rs: PendingApproval serialization
- preview.rs: Text truncation/preview (11 tests)
- skills.rs: Skill selection logic

Shared infrastructure (MockLlmProvider, make_test_agent helpers) lives
in mod.rs. All 47 tests pass.
