+++ src/agent/dispatcher/core.rs
@@ -216,6 +216,12 @@ impl Agent {
         params: &serde_json::Value,
         job_ctx: &JobContext,
     ) -> Result<String, Error> {
-        execute_chat_tool_standalone(self.tools(), self.safety(), tool_name, params, job_ctx).await
+        execute_chat_tool_standalone(
+            self.tools(),
+            self.safety(),
+            &ChatToolRequest { tool_name, params },
+            job_ctx,
+        )
+        .await
     }
 }
